status: Add --booted-index, --pending-index and --rollback-index
Add new commands to output the current, staged and previous deployment for use in automation and scripting.
Right now, it's difficult to pin the current deployment without needing to look into the output of some other tooling (like rpm-ostree) to get the index of each deployment.  This index also is not consistent - the current deployment could be 0 when you first boot the system then 1 shortly after.
This change makes it easy to get the index of the current or future deployment so it can be piped into other commands (such as ostree admin pin).

Huge thanks to Eric for authoring this patch for me!

Co-authored-by: Eric Curtin <ecurtin@redhat.com>
Signed-off-by: Robert Sturla <robertsturla@outlook.com>