Fix booted deploy detection with transient /etc
When /etc is transient, and composefs is root, there is currently no
bind-mount of the deploy directory so the checks for finding if a
deploy is booted are failing.

We fix this by bind-mounting the deploy dir as /run/ostree/deploy and
use that to look for the booted deploy directory.